Move all class to kotlincore package
