+++ app/src/at-vinhnguyen/java/asiantech/internship/summer/kotlincore/exercise1/GreatestCommonDivisorAndLeastCommonMultiple.kt
