+++ app/src/at-vinhnguyen/java/asiantech/internship/summer/kotlincore/exercise11/ListPermutations.kt
